git-svn-id: https://svn.kuali.org/repos/foundation/trunk@8644 eb505253-de9c-4f3b-bf87-3aad84f29c0f

+++ kuali-mvn/src/main/java/org/kuali/maven/common/MvnExecutor.java
@@ -50,7 +50,9 @@ public class MvnExecutor {
             int exitValue = CommandLineUtils.executeCommandLine(cl, stdout, stderr);
             validateExitValue(context, exitValue);
         } finally {
-            deleteQuietly(tempPom);
+            if (context.isDeleteTempPom()) {
+                deleteQuietly(tempPom);
+            }
         }
     }
 
@@ -143,6 +145,8 @@ public class MvnExecutor {
         File file = File.createTempFile("pom.", ".xml", context.getWorkingDir());
         resourceUtils.write(file, s);
         cl.createArg().setValue("-f");
+        log.info("basedir=" + context.getBasedir());
+        log.info("workingDir=" + context.getWorkingDir());
         if (!context.getBasedir().equals(context.getWorkingDir())) {
             File tempPom = new File(context.getBasedir(), file.getName());
             FileUtils.copyFile(file, tempPom);
